+++ Example/ReactComponent/.watchmanconfig
@@ -1,0 +1,1 @@
+{}
